requirements.txt
aiofiles==23.2.1
aiohttp==3.9.1
asyncio==3.4.3
python-dotenv==1.0.0
tinydb==4.8.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
